Fix the cartesian product for higher dimensions
When applying winning line to the fourth dimension on the rules page, the 4D
boards glitched. The reason being the points was in reverse order `[w z y x]`
and not `[x y z w]`.

When picking the circles to fit the points correctly into the svg, we take the 3
first coords in the points vector, since the circles only apply in the third
dimension. When taking the first 3 in reverse order, you get the wrong
duplicates of circles in dimensions above the third. You would never get the
actual necessary coords of `x` and with five dimension `x` and `y`.

The order is therefore imperative.